// src/types/typography.ts
export type TypographyVariant =
  // H1 variants - 36px
  | 'h-bl-36' // H1 Bold 36px
  | 'h-sm-36' // H1 Semibold 36px
  | 'h-rg-36' // H1 Regular 36px
  // H2 variants - 32px
  | 'h2-bl-32' // H2 Bold 32px
  | 'h2-sm-32' // H2 Semibold 32px
  | 'h2-rg-32' // H2 Regular 32px
  // H3 variants - 24px
  | 'h3-bl-24' // H3 Bold 24px
  | 'h3-sm-24' // H3 Semibold 24px
  | 'h3-rg-24' // H3 Regular 24px
  // Body variants - 20px
  | 'body-bl-20' // Body Bold 20px
  | 'body-sm-20' // Body Semibold 20px
  | 'body-rg-20' // Body Regular 20px
  // Subtitle variants - 16px
  | 'subtitle-bl-16' // Subtitle Bold 16px
  | 'subtitle-sm-16' // Subtitle Semibold 16px
  | 'subtitle-rg-16' // Subtitle Regular 16px
  // Caption variants - 14px
  | 'caption-bl-14' // Caption Bold 14px
  | 'caption-sm-14' // Caption Semibold 14px
  | 'caption-rg-14' // Caption Regular 14px
  // Footnote variants - 12px
  | 'footnote-bl-12' // Footnote Bold 12px
  | 'footnote-sm-12' // Footnote Semibold 12px
  | 'footnote-rg-12'; // Footnote Regular 12px

export type TypographyColor =
  | 'primary-black'
  | 'primary-red'
  | 'primary-warm-gray'
  | 'light-gray-1'
  | 'light-gray-2'
  | 'light-gray-3'
  | 'mid-gray-1'
  | 'mid-gray-2'
  | 'mid-gray-3'
  | 'text-1'
  | 'text-2'
  | 'text-3'
  | 'text-4'
  | 'white';